Update stack to Heroku-20 & Python 3.9.x
Setup has been extracted to a bash script so that we can reduce the
number of layers we output, and get a more legible format for editing.

GitHub Actions has been set up to build the container and push it to
DockerHub (when branch is master). I've also added a scheduled build
so our base container will rebuild nightly and keep up to date with
the base images beneath.

Facebook's Watchman has been introduced as a development dependency so
hopefully speed up file watching during local development.