Add keepalive auto-commit
GitHub Actions automatically disables the scheduled workflow runs
if the master branch has not seen any activity in the past 60 days.

This is problematic for this repository as we rely on it to keep our
development Docker images up to date by rebuilding them on a schedule;
combine this with the fact the repository is rarely touched between
major Heroku Stack/Python releases and we hit the problem where we can
silently lose our updating process.

This is the solution: another workflow that commits an empty-commit
to the master branch every 50 days which stops the automatic
de-activation of the docker-build workflow by GitHub.